+++ src/main/java/com/esreven/tickets/controller/EventApiImp.java
@@ -1,9 +1,13 @@
 package com.esreven.tickets.controller;
 
-import com.esreven.tickets.model.EventDto;
-import com.esreven.tickets.service.EventService;
+import com.esreven.tickets.model.EventTicketDto;
+import com.esreven.tickets.model.TicketDto;
+import com.esreven.tickets.model.TicketEventDto;
+import com.esreven.tickets.service.TicketService;
 import io.swagger.annotations.Api;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 import javax.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -16,16 +20,26 @@ import org.springframework.web.bind.annotation.RequestBody;
 @Api(value = "Events", tags={ "events" })
 public class EventApiImp implements EventApi {
 
-    private final EventService eventService;
+    private final TicketService ticketService;
 
-    public EventApiImp(EventService eventService) {
-        this.eventService = eventService;
+    public EventApiImp(TicketService ticketService) {
+        this.ticketService = ticketService;
     }
 
-    public ResponseEntity<EventDto> addEvent(@Valid @RequestBody EventDto body) throws IOException {
+    public ResponseEntity<List<TicketEventDto>> addEvent(@Valid @RequestBody EventTicketDto body)
+        throws IOException {
         log.info("[ POST ] -> /events Body: {}", body);
-        EventDto eventDto = eventService.save(body);
-        return ResponseEntity.status(HttpStatus.CREATED).body(eventDto);
+        List<TicketEventDto> ticketEventDtos = new ArrayList<>();
+        for (TicketDto ticketDto : body.getTicketDto()) {
+            TicketEventDto ticketEventDto = TicketEventDto.builder()
+                .ticketDto(ticketDto)
+                .eventDto(body)
+                .build();
+
+            ticketEventDtos.add(ticketService.save(ticketEventDto));
+        }
+
+        return ResponseEntity.status(HttpStatus.CREATED).body(ticketEventDtos);
     }
 
 }
