+++ src/main/java/com/esreven/tickets/model/EventDto.java
@@ -1,6 +1,5 @@
 package com.esreven.tickets.model;
 
-import com.esreven.tickets.domain.Ticket;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import javax.validation.constraints.NotNull;
 import lombok.Builder;
@@ -10,34 +9,28 @@ import org.springframework.validation.annotation.Validated;
 @Data
 @Builder
 @Validated
-public class EventDto extends  {
+public class EventDto {
 
     @JsonProperty("eventId")
     @NotNull
     private Long eventId;
 
     @JsonProperty("totalListings")
-    @NotNull
     private Long totalListings;
 
     @JsonProperty("totalTickets")
-    @NotNull
     private Long totalTickets;
 
     @JsonProperty("minQuantity")
-    @NotNull
     private int minQuantity;
 
     @JsonProperty("maxQuantity")
-    @NotNull
     private int maxQuantity;
 
     @JsonProperty("start")
-    @NotNull
     private int start;
 
     @JsonProperty("rows")
-    @NotNull
     private int rows;
 
 }
