--- src/main/java/com/esreven/tickets/model/EventTicketsDto.java
@@ -1,12 +1,0 @@
-package com.esreven.tickets.model;
-
-import lombok.Builder;
-import lombok.Data;
-import org.springframework.validation.annotation.Validated;
-
-@Data
-@Builder
-@Validated
-public class EventTicketsDto extends Dto {
-
-}
