--- src/main/java/com/esreven/tickets/service/EventService.java
@@ -1,9 +1,0 @@
-package com.esreven.tickets.service;
-
-import com.esreven.tickets.model.EventDto;
-import java.io.IOException;
-
-public interface EventService {
-
-    EventDto save(EventDto eventDto) throws IOException;
-}
